Using a more robust type

+++ lista02/q01/include/galois_arithmetic.h
@@ -1,8 +1,10 @@
 #ifndef GALOIS_ARITHMETIC_H
 #define GALOIS_ARITHMETIC_H
 
-long gf_xor(long a, long b);
-long gf_mult(long a, long b);
-long gf_div(long a, long b);
+#include <stdint.h>
+
+uint8_t gf_xor(uint8_t a, uint8_t b);
+uint8_t gf_mult(uint8_t a, uint8_t b);
+uint8_t gf_div(uint8_t a, uint8_t b);
 
 #endif
